saves 6k more ram
